refactor: 의존성 분리
