+++ React/util/typeGuard.ts
